backend/requirements.txt
django
djangorestframework
django-environ
psycopg2-binary
django-cors-headers
dj-database-url
gunicorn
whitenoise
django-filter
google-genai
Pillow
djangorestframework-simplejwt
django-allauth
dj-rest-auth